Development
Added CheatActivity with layout file.  A user can select to cheat and the activity launches.  They can then press the button to show the answer.
Next step is to hold the user accountable when they go back and answer the question.

Stopped at figure 5.12 on page 101
Changes to be committed:
    modified:   app/src/main/java/com/example/android/betrueornotbetrue/MainActivity.java
    modified:   app/src/main/res/layout/activity_main.xml