Development
Finished up chapter 5.  Now when a user selects to show the answer in the cheat activity MainActivity is aware of it because the data is passed back.
There are three loopholes to fix: Users can roate CheatActivity to reset and erase thier cheat, they can rotate MainActivity to clear out the mIsCheater variable when returning to the view, and the can press the next button and go all the way around until they get back and then answer the question and not be shown the cheat message.

Stopped at pg 109.  Finish the chapter 5 challenges.
Changes to be committed:
    modified:   app/src/main/java/com/example/android/betrueornotbetrue/MainActivity.java