Development
Added in animation code from a later API.  this was part of section, "Adding code from later APIs saftely" in chapter 6.  The lesson is to run an if statement checking the build version for the later API code, if the build verion matches what is needed you run the code, if not then set a fallback.

Changes to be committed:
    modified:   .idea/workspace.xml
    modified:   app/src/main/java/com/example/android/betrueornotbetrue/CheatActivity.java